+++ explore/shared/src/main/scala/doodle/explore/syntax/ExploreSyntax.scala
@@ -28,16 +28,24 @@ import monix.execution.Scheduler
 trait ExploreSyntax {
   implicit class ExploreFunctionOps[A, Alg[x[_]] <: Algebra[x], F[_], B](
       f: A => Picture[Alg, F, B]) {
-    def explore[Frame, Canvas]()(implicit ex: ExplorerFactory[_, A],
+    def explore[Frame, Canvas](frame: Frame)(implicit ex: ExplorerFactory[_, A],
                                  a: AnimationRenderer[Canvas],
                                  e: DefaultRenderer[Alg, F, Frame, Canvas],
                                  s: Scheduler,
-                                 m: Monoid[B]): B = {
+                                 m: Monoid[B]): Unit = {
       (for {
-        canvas <- e.canvas(e.default)
+        canvas <- e.canvas(frame)
         values <- ex.create.render
         b <- a.animate(canvas)(values.map(f))
-      } yield b).unsafeRunSync()
+       } yield b).unsafeRunAsync(r =>
+        r match {
+          case Left(err) =>
+            println("There was an error rendering an animation")
+            err.printStackTrace()
+
+          case Right(_) => ()
+        }
+      )
     }
   }
 }