+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.9.14
+## 0.9.15
+
+- `explore` takes a `Frame`, which means it now works again.
+
+
+## 0.9.14 20-Dec-2019
 
 - `Picture` has a `Monad` instance if the underlying type `F` does.
 
@@ -13,6 +18,8 @@
 - Add `Debug` algebra, which draws bounding box and origin for the given picture.
 
 - Draw bitmaps the correct way around (they were upside-down).
+
+- Add some easing functions, the start of a library to make animations easier.
 
 - Miscellaneous small improvements.